requirements.txt
colorama==0.4.1
confusable-homoglyphs==3.2.0
Django==2.1.7
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-db-signals==0.1.1
django-filter==2.1.0
django-js-asset==1.2.2
django-registration==3.0
django-wysiwyg==0.8.0
mccabe==0.6.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
